feat(events): add child object as parameter to child events (for lv_obj_set_parent and lv_obj_move_to_index)
Some events that modify the children of objects send an event to the parent to inform them that their child has changed. The child is passed as an event parameter so that the parent can use the knowledge about which child was impacted.
This commit adds the child as parameter to lv_obj_set_parent and lv_obj_move_to_index.